+++ Assets/Prefab/Enemies/Enemy_Plant.prefab
@@ -28,7 +28,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3219288358420711660}
   serializedVersion: 2
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 6.28, y: -2.56, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -138,7 +138,7 @@ MonoBehaviour:
   bulletPrefab: {fileID: 5111563969775021251, guid: fb2219765cbc2b3478872eab470e43f4, type: 3}
   bulletOrigin: {fileID: 3219288359241729508}
   bulletSpeed: 12
-  facingRight: 0
+  facingRight: 1
 --- !u!61 &3219288358420711659
 BoxCollider2D:
   m_ObjectHideFlags: 0
