+++ Assets/Scripts/Traps/Danger.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Danger : MonoBehaviour
+{
+    // this one responsible for giving damage to the player
+
+    protected virtual void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.GetComponent<Player>() != null)
+        {
+            Player player = collision.GetComponent<Player>();
+
+            player.Knockback(transform);
+        }
+            
+    }
+}
