+++ Assets/Scripts/Traps/Danger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb48b1e05b144a44dbf53c665d6396f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
