+++ Library/EditorInstance.json
@@ -1,0 +1,6 @@
+{
+	"process_id" : 11820,
+	"version" : "2022.3.8f1",
+	"app_path" : "D:/Programs/Unity/2022.3.8f1/Editor/Unity.exe",
+	"app_contents_path" : "D:/Programs/Unity/2022.3.8f1/Editor/Data"
+}
