+++ Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,0 +1,3 @@
+Base path: 'D:/Programs/Unity/2022.3.8f1/Editor/Data', plugins path 'D:/Programs/Unity/2022.3.8f1/Editor/Data/PlaybackEngines'
+Cmd: initializeCompiler
+
