+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2022.3.13f1
-m_EditorVersionWithRevision: 2022.3.13f1 (5f90a5ebde0f)
+m_EditorVersion: 2022.3.8f1
+m_EditorVersionWithRevision: 2022.3.8f1 (b5eafc012955)
