+++ UserSettings/EditorUserSettings.asset
@@ -6,31 +6,31 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 0009505406510a0b54570873412708444e164d7f287f2032287c4e32e7b36260
+      value: 5000550551040d5d085d09774021594415164c72782c7167287a4463e1b0366e
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5a09020403005f03545a5c2340740b44174f4b7f787177357f7e4b61b1b46061
+      value: 5200020700060a0d0f585c7313735c44434f1e2b782c253178701867b6e3626e
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 0155040357515b5f585f092447250f44174f4e7a747d226128784436e0b7656a
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 5457520554535f0d0f0d0a2716710744444e40282d7c77627b2c1960bbe4603e
-      flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 5552075301505b5a095f5877162706444e151e7a2d7c706574791865e6b56d6a
-      flags: 0
-    RecentlyUsedSceneGuid-5:
-      value: 5300075253060d095e5a5b7216705a444f4e1b282d717f682c2a1b60b2e46369
-      flags: 0
-    RecentlyUsedSceneGuid-6:
-      value: 5155030707065b0d0f580d2413740e4441151a292e2e74362e7c1e67e7e1676e
-      flags: 0
-    RecentlyUsedSceneGuid-7:
       value: 5205555300530d0b0e5f0f2043760b44104f49297e7873677f2c4d35b7b0373c
       flags: 0
+    RecentlyUsedSceneGuid-3:
+      value: 5155030707065b0d0f580d2413740e4441151a292e2e74362e7c1e67e7e1676e
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 5300075253060d095e5a5b7216705a444f4e1b282d717f682c2a1b60b2e46369
+      flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 5552075301505b5a095f5877162706444e151e7a2d7c706574791865e6b56d6a
+      flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 0257070203540d0a0e0b0d77142507444f15197d782972352c784d63b6b1316f
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 5457520554535f0d0f0d0a2716710744444e40282d7c77627b2c1960bbe4603e
+      flags: 0
     RecentlyUsedSceneGuid-8:
-      value: 5200020700060a0d0f585c7313735c44434f1e2b782c253178701867b6e3626e
+      value: 5302045f54070d5f0859542313275d4441164f7e7c797f36782c1b66b1b0616b
       flags: 0
     RecentlyUsedSceneGuid-9:
       value: 51500205575059035c595576117a5c44434f4d2f2d7b27627f791831bbe66061
